refactor: Rename run.dto.ts
