+++ server/tests/e2e/session.e2e-spec.ts
@@ -9,7 +9,7 @@ import { del, post } from '../util/request-handlers.util';
 import { RunSessionDto } from '@common/dto/run/run-session.dto';
 import { RunSessionTimestampDto } from '@common/dto/run/run-session-timestamp.dto';
 import { UserMapRankDto } from '@common/dto/run/user-map-rank.dto';
-import { RunDto } from '@common/dto/run/runs.dto';
+import { RunDto } from '@common/dto/run/run.dto';
 import { XpSystemsService } from '@modules/xp-systems/xp-systems.service';
 import { RunValidationErrorTypes } from '@common/enums/run.enum';
